requirements.txt
torchaudio
transformers
gradio
peft
accelerate
